fix(policy): holistic review findings (circular dep, dead re-check, typing, threat doc, perf TODO)
H-1: Break circular dep between pipeline/runner and policy/path-gate by
extracting matchesAnyPattern into util/glob.ts.

H-2: Remove dead cost-gate re-check after the initial evaluatePolicyGates
call — it used the same cumulative tokensUsed snapshot so it could never
observe a different value. Relabel the initial gate stage as "all-stages"
since tokensUsed at that point is cumulative across the pipeline.

M-1: Replace `this.db as any` with `this.db as AnyDb` at the policy
reviewers-requested audit site for consistency.

M-2: Document the override-label bypass threat model on hasOverrideLabel.

M-3: Add perf TODO at verifyApprovalsReceived call site noting the need
for per-(org,team) caching to avoid secondary rate limits.
