+++ packages/core/src/__tests__/auto-merge.test.ts
@@ -16,7 +16,8 @@ import { describe, it, expect, vi, beforeEach } from "vitest";
 import { createHmac } from "node:crypto";
 import { Hono } from "hono";
 import { createDb } from "../db/client.js";
-import { PipelineRunner, matchesAnyPattern } from "../pipeline/runner.js";
+import { PipelineRunner } from "../pipeline/runner.js";
+import { matchesAnyPattern } from "../util/glob.js";
 import { createWebhookHandler } from "../webhook/handler.js";
 import { pipelineRuns } from "../db/schema.js";
 import { eq } from "drizzle-orm";
